+++ src/compiler/index.ts
@@ -14,6 +14,7 @@ export * from "./analysis/visitors/SemanticValidatorAstVisitor";
 export * from "./analysis/visitors/SyntaxRulesAstVisitor";
 
 export * from "./transformation/visitors/DesugarAstVisitor";
+export * from "./transformation/visitors/ComptimeEvaluationAstVisitor";
 export * from "./transformation/visitors/InlineImportsAstVisitor";
 export * from "./transformation/visitors/TreeShakeAstVisitor";